requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2
python-dotenv  # (if using .env files)
